moving storybook to root website
