--- examples/basic-example/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en-us">
-<head>
-    <meta charset="UTF-8">
-    <title>React Sortable Tree</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="apple-touch-icon" sizes="180x180" href="static/apple-touch-icon.png">
-    <link rel="icon" type="image/png" href="static/favicon-32x32.png" sizes="32x32">
-    <link rel="icon" type="image/png" href="static/favicon-16x16.png" sizes="16x16">
-    <link rel="mask-icon" href="static/safari-pinned-tab.svg" color="#de7a32">
-    <link rel="shortcut icon" href="static/favicon.ico">
-    <meta name="theme-color" content="#ffffff">
-</head>
-<body>
-    <div id="app"></div>
-</body>
-</html>
