--- examples/basic-example/stylesheets/app.css
@@ -1,25 +1,0 @@
-.page-header {
-  padding: 1rem 6rem;
-}
-
-.main-content {
-  box-sizing: border-box;
-}
-
-body {
-  background-color: #fafafa;
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-/**
- * Uncomment to simulate bootstrap environment
- */
-/*
-html {
-    box-sizing: border-box;
-}
-
-*, *::before, *::after {
-    box-sizing: inherit;
-}
-*/
