+++ package.json
@@ -4,12 +4,12 @@
   "description": "Drag-and-drop sortable component for nested data and hierarchies",
   "scripts": {
     "prebuild": "npm run lint",
-    "build": "npm run build:commonjs && npm run build:es && npm run build:umd && npm run build:demo",
+    "build": "npm run build:commonjs && npm run build:es && npm run build:umd && npm run build:storybook",
     "build:umd": "npm run clean:umd && cross-env NODE_ENV=rollup rollup -c",
     "build:es": "npm run clean:es && cross-env NODE_ENV=es babel src --out-dir dist/es",
     "build:commonjs": "npm run clean:commonjs && cross-env NODE_ENV=commonjs babel src --out-dir dist/commonjs",
-    "build:demo": "npm run clean:demo && cross-env NODE_ENV=production webpack --bail && npm run build-storybook",
-    "clean:demo": "rimraf build",
+    "build:storybook": "npm run clean:storybook && build-storybook -o build",
+    "clean:storybook": "rimraf build",
     "clean:commonjs": "rimraf dist/commonjs",
     "clean:es": "rimraf dist/es",
     "clean:umd": "rimraf dist/umd",
@@ -20,9 +20,8 @@
     "release": "standard-version",
     "test": "cross-env NODE_ENV=test jest",
     "test:watch": "cross-env NODE_ENV=test jest --watchAll",
-    "deploy": "npm run build:demo && gh-pages -d build",
-    "storybook": "start-storybook -p ${PORT:-3001} -h 0.0.0.0",
-    "build-storybook": "build-storybook -o build/storybook"
+    "deploy": "npm run build:storybook && gh-pages -d build",
+    "storybook": "start-storybook -p ${PORT:-3001} -h 0.0.0.0"
   },
   "main": "dist/commonjs/index.js",
   "module": "dist/es/index.js",
