+++ stories/rtl-support.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import SortableTree from '../../src';
+import SortableTree from '../src';
 // In your own app, you would need to use import styles once in the app
 // import 'react-sortable-tree/styles.css';
 
@@ -9,15 +9,27 @@ export default class App extends Component {
 
     this.state = {
       treeData: [
-        { title: 'Chicken', expanded: true, children: [{ title: 'Egg' }] },
+        {
+          title: 'Chicken',
+          expanded: true,
+          children: [
+            { title: 'Egg' },
+            { title: 'Egg' },
+            { title: 'Egg' },
+            { title: 'Egg' },
+            { title: 'Egg' },
+            { title: 'Egg' },
+          ],
+        },
       ],
     };
   }
 
   render() {
     return (
-      <div style={{ height: 300 }}>
+      <div style={{ height: 300, width: 600 }}>
         <SortableTree
+          rowDirection="rtl"
           treeData={this.state.treeData}
           onChange={treeData => this.setState({ treeData })}
         />
