+++ stories/search.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import SortableTree from '../../src';
+import SortableTree from '../src';
 // In your own app, you would need to use import styles once in the app
 // import 'react-sortable-tree/styles.css';
