+++ stories/tree-data-io.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import SortableTree, {
   getFlatDataFromTree,
   getTreeFromFlatData,
-} from '../../src';
+} from '../src';
 // In your own app, you would need to use import styles once in the app
 // import 'react-sortable-tree/styles.css';
